[CHORE] Merge 충돌 해결

+++ src/main/java/com/server/youthtalktalk/domain/policy/dto/ReviewInPolicyDto.java
@@ -6,8 +6,8 @@ public record ReviewInPolicyDto(
         Long postId,
         String title,
         String contentPreview,
-        int commentCount,
-        int scrapCount,
+        long commentCount,
+        long scrapCount,
         LocalDate createdAt
 ) {
 }
